feat: Basic functionalities add, remove and list Items.
This Commit contains the first set of functionalities.
The overall Idea is that we store the current stock of the pantry.
The App is using the barcode to get informations from the Item. The
barcode is also the general id for this Item. I thought this is the best
way. The barcode is unique (as far as I know). Thanks to this the
barcode can also used to add and delete items.

From this barcode we got also the information about the package size
(units). That means you can add the number of packages to the stock.
With the package information we can store both amount and package
number. Currently we can only add whole packages.
When you want to delete items you can handover packages or
amount. The new stock will calculated.

This code is completely from chatgpt. I just made smaller adjustments.